refactor: extract :app composition root and enforce clean architecture
- Add thin :app module as the Hilt/Android entry point (@HiltAndroidApp);
  presentation becomes a pure Android library with no applicationId
- Delete :infra module; move RepositoryModule to :data (using @Binds) and
  add @Inject constructors to all repository implementations
- Remove hilt-android from :domain; replace with javax.inject + dagger-core
  so domain stays framework-agnostic while KSP-generated factories still compile
- Add kotlinx-coroutines-core explicitly to :domain (was previously transitive via hilt-android)
- Remove manual AppDatabase singleton; fully delegate lifecycle to Hilt
- Strip app-level manifest attributes from :presentation manifest (merged at build time)
