+++ app/build.gradle
@@ -1,17 +1,19 @@
 plugins {
-    alias(libs.plugins.android.library)
+    alias(libs.plugins.android.application)
+    alias(libs.plugins.hilt)
     alias(libs.plugins.ksp)
 }
 
 android {
     compileSdk 35
-
     defaultConfig {
+        applicationId "com.anwera64.pagodividido"
         minSdk 21
         targetSdk 35
-        consumerProguardFiles "consumer-rules.pro"
+        versionCode 1
+        versionName "1.0"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
     }
-
     buildTypes {
         release {
             minifyEnabled false
@@ -22,19 +24,14 @@ android {
         sourceCompatibility JavaVersion.VERSION_17
         targetCompatibility JavaVersion.VERSION_17
     }
-    namespace 'com.anwera97.domain'
+    namespace 'com.anwera64.pagodividido.app'
 }
 
 dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation project(':presentation')
+    implementation project(':data')
+    implementation project(':domain')
 
-    implementation libs.javax.inject
-    implementation libs.dagger
-    implementation libs.kotlinx.coroutines.core
+    implementation libs.hilt.android
     ksp libs.hilt.compiler
-
-    testImplementation libs.junit
-    testImplementation libs.mockito.core
-    testImplementation libs.mockito.kotlin
-    testImplementation libs.kotlinx.coroutines.test
 }
