+++ app/src/main/java/com/anwera64/pagodividido/app/PagoDividioApp.kt
@@ -1,4 +1,4 @@
-package com.anwera64.pagodividido
+package com.anwera64.pagodividido.app
 
 import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
