+++ gradle/libs.versions.toml
@@ -20,6 +20,7 @@ androidx-test-junit     = "1.2.1"
 espresso                = "3.6.1"
 mockito-core            = "5.12.0"
 mockito-kotlin          = "5.4.0"
+javax-inject            = "1"
 
 [libraries]
 androidx-core-ktx                       = { module = "androidx.core:core-ktx",                             version.ref = "core-ktx" }
@@ -49,6 +50,7 @@ androidx-room-runtime                   = { module = "androidx.room:room-runtime
 androidx-room-compiler                  = { module = "androidx.room:room-compiler",                       version.ref = "room" }
 androidx-room-ktx                       = { module = "androidx.room:room-ktx",                            version.ref = "room" }
 androidx-room-testing                   = { module = "androidx.room:room-testing",                        version.ref = "room" }
+dagger                                  = { module = "com.google.dagger:dagger",                         version.ref = "hilt" }
 hilt-android                            = { module = "com.google.dagger:hilt-android",                    version.ref = "hilt" }
 hilt-compiler                           = { module = "com.google.dagger:hilt-compiler",                   version.ref = "hilt" }
 kotlinx-coroutines-core                 = { module = "org.jetbrains.kotlinx:kotlinx-coroutines-core",     version.ref = "coroutines" }
@@ -60,6 +62,7 @@ androidx-test-espresso-core             = { module = "androidx.test.espresso:esp
 androidx-arch-core-testing              = { module = "androidx.arch.core:core-testing",                   version.ref = "arch-core" }
 mockito-core                            = { module = "org.mockito:mockito-core",                           version.ref = "mockito-core" }
 mockito-kotlin                          = { module = "org.mockito.kotlin:mockito-kotlin",                  version.ref = "mockito-kotlin" }
+javax-inject                            = { module = "javax.inject:javax.inject",                          version.ref = "javax-inject" }
 
 [bundles]
 room        = ["androidx-room-runtime", "androidx-room-ktx"]
