+++ presentation/build.gradle
@@ -1,5 +1,5 @@
 plugins {
-    alias(libs.plugins.android.application)
+    alias(libs.plugins.android.library)
     alias(libs.plugins.hilt)
     alias(libs.plugins.kotlin.compose)
     alias(libs.plugins.ksp)
@@ -8,13 +8,8 @@ plugins {
 android {
     compileSdk 35
     defaultConfig {
-        applicationId "com.anwera64.pagodividido"
         minSdk 21
         targetSdk 35
-        versionCode 1
-        versionName "1.0"
-        vectorDrawables.useSupportLibrary = true
-        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
     }
     buildTypes {
         release {
@@ -37,7 +32,6 @@ android {
 
 dependencies {
     implementation project(':domain')
-    implementation project(':infra')
     implementation fileTree(dir: 'libs', include: ['*.jar'])
 
     implementation libs.material
