+++ settings.gradle
@@ -17,7 +17,7 @@ dependencyResolutionManagement {
     }
 }
 
+include ':app'
 include ':presentation'
 include ':data'
 include ':domain'
-include ':infra'
